+++ NEWS.md
@@ -3,6 +3,7 @@
 * Functions get_trait_data() and get_taxon_data() now return 'OriginalReference' and 'OriginalDOI' if available 
 * New trait definitions 'StemEPS', 'CCleaf', 'CCsapwood' and 'CCfineroot' and leaf phenology traits
 * New functions 'check_harmonized_trait_dir()' and 'check_harmonized_allometry_dir()'
+* Trait data can be harmonized in wide and long format
 
 # traits4models 0.2.0
 * New columns 'MininumValue' and 'MaximumValue' added to HarmonizedTraitDefinition